Update README.md
Provided a structure to the readme. We should discuss more on what we would like to include in the README but I believe I got most of it for now. If you are able to add something to the readme, please do! I believe that it is important we document what we can whenever we can.