Documented the course registry use case
Added Javadoc to the course register use case files. Some changes still need to be made within the Javadoc, specifically the prepareFailView in the RegisterPresenter.

I believe some fixes should be made with the naming of the CRegisterPresenter as I believe it acts as an output boundary and should be named as such.