Update documentation for the vote on solution document use case
Added documentation for the vote on solution use case.

Note that some parts of this is still incomplete which is reflected by todos in the javadoc.

Noticed a small error with the output boundary and presenter parameters and after discussing with @pandaponn decided to just quickly change it within this PR