Select different solutions
Added a dropdown menu on the SolutionFrame that allows to switch between 3 different solutions. This is currently hard-coded and will be replaced with whatever solutions are available within the src/data directory.

Addition of a data directory which will hold the pdf files that the ui will access.

Updated the test toolbar to create a solution window when solutions button is pressed. Although the button should be disabled until a test has been completed, I decided to enable it for testing purposes.